# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: AIPC Platform API
  description: 'AIPC Platform REST API for edge AI computing platform.


    Provides unified access to:

    - AI Runtime (model management, inference)

    - Event Bus (pub/sub messaging)

    - Device Control (camera, PTZ, GPIO)

    - App Manager (container lifecycle)

    - Stream (HLS video streaming)

    - Error semantics (all endpoints): validation failures return HTTP 400
      with envelope {code: 1001, message: "Invalid request", error: {detail}};
      a missing, invalid or revoked Bearer token returns HTTP 401 with
      code 2000. Operations below document their success responses; the
      400/401 error behaviour applies to every write endpoint even where
      not repeated per-operation.

    '
  version: 1.0.2
  contact:
    name: AIPC Platform Team
    email: support@camthink.ai
servers:
- url: https://{host}/api/v1
  description: Production — nginx TLS gateway on port 443 (self-signed
    certificate, disable certificate verification in clients). The only
    base reachable from remote machines.
  variables:
    host:
      default: localhost
- url: /api/v1
  description: Current host (resolved against wherever this spec is served,
    e.g. http://<device>:8080/api/v1 — same-origin, no CORS)
- url: http://localhost:8080/api/v1
  description: Local debugging on the device only — platform-api binds
    127.0.0.1:8080, the port is not reachable remotely
security:
- BearerAuth: []
...
paths:
...
  /containers:
    get:
      tags: [Containers]
      summary: List containers
      operationId: listContainers
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [running, stopped, all]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of containers
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        containers:
                          type: array
                          items:
                            $ref: '#/components/schemas/Container'
                        total:
                          type: integer
                        status_count:
                          type: object
  /containers/{id}:
    get:
      tags: [Containers]
      summary: Get container details
      operationId: getContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Container details
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        container:
                          $ref: '#/components/schemas/Container'
    delete:
      tags: [Containers]
      summary: Remove container
      operationId: removeContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Container removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
  /containers/{id}/stats:
    get:
      tags: [Containers]
      summary: Get container stats
      operationId: getContainerStats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Container stats
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        stats:
                          $ref: '#/components/schemas/ContainerStats'
  /containers/{id}/logs:
    get:
      tags: [Containers]
      summary: Get container logs
      operationId: getContainerLogs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tail
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Container logs
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        logs:
                          type: string
  /containers/{id}/logs/stream:
    get:
      tags: [Containers]
      summary: Stream container logs (SSE)
      operationId: streamContainerLogs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Log stream
          content:
            text/event-stream:
              schema:
                type: string
  /containers/{id}/start:
    post:
      tags: [Containers]
      summary: Start container
      operationId: startContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Container started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
  /containers/{id}/stop:
    post:
      tags: [Containers]
      summary: Stop container
      operationId: stopContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Container stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
  /containers/{id}/restart:
    post:
      tags: [Containers]
      summary: Restart container
      operationId: restartContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Container restarted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  # ==================== Images API ====================
...
  /images/pull:
    post:
      tags: [Images]
      summary: Pull an image
      operationId: pullImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
      responses:
        '200':
          description: Image pulled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
  /images/{image}:
    delete:
      tags: [Images]
      summary: Delete an image
      operationId: deleteImage
      parameters:
        - name: image
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  # ==================== App Manager Extensions ====================
...
  /containers/{id}/logs/ws:
    get:
      tags: [App Manager]
      summary: Stream container logs in real-time over WebSocket
      operationId: streamContainerLogsWS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tail
          in: query
          required: false
          schema:
            type: integer
            default: 100
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        '101':
          description: Switching to WebSocket protocol
        '400':
          description: Invalid request
  /containers/{id}/exec/ws:
    get:
      tags: [App Manager]
      summary: Interactive container console over WebSocket
      operationId: execContainerWS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: cols
          in: query
          required: false
          schema:
            type: integer
            default: 80
        - name: rows
          in: query
          required: false
          schema:
            type: integer
            default: 24
        - name: command
          in: query
          required: false
          schema:
            type: string
            default: "/bin/sh"
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        '101':
          description: Switching to WebSocket protocol
        '400':
          description: Invalid request

  # ==================== Storage Management ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    APIResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: Success
        data:
          type: object
          description: Response payload
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorDetail:
      type: object
      properties:
        type:
          type: string
        detail:
          type: string

    # ==================== Store Schemas ====================
...
    Container:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
        state:
          type: string
        created:
          type: string
          format: date-time
        labels:
          type: object
          additionalProperties:
            type: string
    ContainerStats:
      type: object
      properties:
        cpu_percent:
          type: number
        memory_usage:
          type: integer
          format: int64
        memory_limit:
          type: integer
          format: int64
        memory_percent:
          type: number
        network_rx:
          type: integer
          format: int64
        network_tx:
          type: integer
          format: int64
        pids:
          type: integer
...